calimero_runtime::logic
Struct VMHostFunctions

pub struct VMHostFunctions<'a> { /* private fields */ }

The self-referencing struct.

Implementations

impl VMHostFunctions<'_>

pub fn panic( &self, file_ptr: u64, file_len: u64, line: u32, column: u32, ) -> VMLogicResult<()>

pub fn panic_utf8( &self, msg_ptr: u64, msg_len: u64, file_ptr: u64, file_len: u64, line: u32, column: u32, ) -> VMLogicResult<()>

pub fn register_len(&self, register_id: u64) -> VMLogicResult<u64>

pub fn read_register( &self, register_id: u64, ptr: u64, len: u64, ) -> VMLogicResult<u32>

pub fn context_id(&mut self, register_id: u64) -> VMLogicResult<()>

pub fn executor_id(&mut self, register_id: u64) -> VMLogicResult<()>

pub fn input(&mut self, register_id: u64) -> VMLogicResult<()>

pub fn value_return( &mut self, tag: u64, ptr: u64, len: u64, ) -> VMLogicResult<()>

pub fn log_utf8(&mut self, ptr: u64, len: u64) -> VMLogicResult<()>

pub fn emit( &mut self, kind_ptr: u64, kind_len: u64, data_ptr: u64, data_len: u64, ) -> VMLogicResult<()>

pub fn commit( &mut self, root_hash_ptr: u64, root_hash_len: u64, artifact_ptr: u64, artifact_len: u64, ) -> VMLogicResult<()>

pub fn storage_read( &mut self, key_ptr: u64, key_len: u64, register_id: u64, ) -> VMLogicResult<u32>

pub fn storage_remove( &mut self, key_ptr: u64, key_len: u64, register_id: u64, ) -> VMLogicResult<u32>

pub fn storage_write( &mut self, key_ptr: u64, key_len: u64, value_ptr: u64, value_len: u64, register_id: u64, ) -> VMLogicResult<u32>

pub fn fetch( &mut self, url_ptr: u64, url_len: u64, method_ptr: u64, method_len: u64, headers_ptr: u64, headers_len: u64, body_ptr: u64, body_len: u64, register_id: u64, ) -> VMLogicResult<u32>

pub fn random_bytes(&mut self, ptr: u64, len: u64) -> VMLogicResult<()>

pub fn time_now(&mut self, ptr: u64, len: u64) -> VMLogicResult<()>

Gets the current time.

This function obtains the current time as a nanosecond timestamp, as SystemTime is not available inside the guest runtime. Therefore the guest needs to request this from the host.

pub fn send_proposal( &mut self, actions_ptr: u64, actions_len: u64, id_ptr: u64, id_len: u64, ) -> VMLogicResult<()>

Call the contract’s send_proposal() function through the bridge.

The proposal actions are obtained as raw data and pushed onto a list of proposals to be sent to the host.

Note that multiple actions are received, and the entire batch is pushed onto the proposal list to represent one proposal.

Parameters

actions_ptr - Pointer to the start of the action data in WASM memory.

actions_len - Length of the action data.

id_ptr - Pointer to the start of the id data in WASM memory.

id_len - Length of the action data. This should be 32 bytes.

pub fn approve_proposal( &mut self, approval_ptr: u64, approval_len: u64, ) -> VMLogicResult<()>